Relocated constants and migrated to registry based digester
