+++ packages/cydran/src/behavior/core/MultiSelectValueModelBehavior.ts
@@ -1,7 +1,6 @@
 import AbstractBehavior from "behavior/AbstractBehavior";
 import Attrs from "const/AttrsFields";
-import { INPUT_KEY, DOM_KEY } from "Constants";
-import { BEHAVIOR_FORM_RESET, CHANGE_KEY } from "const/HardValues";
+import { INPUT_KEY, DOM_KEY, BEHAVIOR_FORM_RESET, CHANGE_KEY } from "Constants";
 
 class MultiSelectValueModel extends AbstractBehavior<string | string[], HTMLSelectElement, any> {
