+++ packages/cydran/src/component/renderer/StageRendererImpl.ts
@@ -3,7 +3,7 @@ import ComponentIdPair from "component/CompnentIdPair";
 import { SelectorError } from "error/Errors";
 import Attrs from "const/AttrsFields";
 import TagNames from "const/TagNames";
-import { ATTRIBUTE_DELIMITER, DEFAULT_PREFIX } from "const/HardValues";
+import { ATTRIBUTE_DELIMITER, DEFAULT_PREFIX } from "Constants";
 import DomUtils from "dom/DomUtils";
 
 const CYDRAN_PREFIX: string = DEFAULT_PREFIX + ATTRIBUTE_DELIMITER;
