--- packages/cydran/src/const/IdsFields.ts
@@ -1,5 +1,0 @@
-enum Ids {
-	STAGE = "Stage"
-}
-
-export default Ids;
