--- packages/cydran/src/const/PropertyKeys.ts
@@ -1,20 +1,0 @@
-enum PropertyKeys {
-	CYDRAN_DIGEST_MAX_EVALUATIONS = "cydran.digest.maxEvaluations",
-	CYDRAN_CLONE_MAX_EVALUATIONS = "cydran.clone.maxEvaluations",
-	CYDRAN_EQUALS_MAX_EVALUATIONS = "cydran.equals.maxEvaluations",
-	CYDRAN_STARTUP_SYNCHRONOUS = "cydran.startup.synchronous",
-	CYDRAN_STRICT_ENABLED = "cydran.strict.enabled",
-	CYDRAN_STRICT_STARTPHRASE = "cydran.strict.startphrase",
-	CYDRAN_LAZY_STARTPHRASE = "cydran.lazy.startphrase",
-	CYDRAN_STRICT_MESSAGE = "cydran.strict.message",
-	CYDRAN_LOG_LEVEL = "cydran.logging.level",
-	CYDRAN_STYLES_ENABLED = "cydran.styles.enabled",
-	CYDRAN_OVERRIDE_WINDOW = "cydran.override.window",
-	CYDRAN_LOG_STRATEGY = "cydran.logging.strategy",
-	CYDRAN_LOG_COLOR_PREFIX = "cydran.logging.color",
-	CYDRAN_LOG_LABEL = "cydran.logging.label",
-	CYDRAN_LOG_LABEL_VISIBLE = "cydran.logging.label.visible",
-	CYDRAN_LOG_PREAMBLE_ORDER = "cydran.logging.pramble.order"
-}
-
-export default PropertyKeys;
