+++ packages/cydran/src/digest/SegmentDigesterImpl.ts
@@ -1,27 +1,14 @@
 import DigestionCandidate from "digest/DigestionCandidate";
 import SegmentDigester from "digest/SegmentDigester";
 import Logger from "log/Logger";
-import LoggerFactory from "log/LoggerFactory";
-import { extractClassName, isDefined } from "util/Utils";
-
-const DEFAULT_FACTORY: () => SegmentDigesterImpl = () => new SegmentDigesterImpl();
+import { extractClassName, requireNotNull } from 'util/Utils';
 
 class SegmentDigesterImpl implements SegmentDigester {
 
 	private logger: Logger;
 
-	private static factory: () => SegmentDigesterImpl = DEFAULT_FACTORY;
-
-	constructor() {
-		this.logger = LoggerFactory.getLogger("SegmentDigester");
-	}
-
-	public static create(): SegmentDigesterImpl {
-		return SegmentDigesterImpl.factory();
-	}
-
-	public static setFactory(factory: () => SegmentDigesterImpl): void {
-		SegmentDigesterImpl.factory = isDefined(factory) ? factory : DEFAULT_FACTORY;
+	constructor(logger: Logger) {
+		this.logger = requireNotNull(logger, "logger");
 	}
 
 	public digestSegment(id: string, changedCandidates: DigestionCandidate[], candidates: DigestionCandidate[]): void {
