+++ packages/cydran/src/index.ts
@@ -1,4 +1,4 @@
-import { CYDRAN_KEY, Ids } from "Constants";
+import { CYDRAN_KEY, PropertyKeys } from "Constants";
 import ArgumentsResolvers from 'argument/ArgumentsResolvers';
 import AbstractBehavior from "behavior/AbstractBehavior";
 import AbstractValueBehavior from "behavior/AbstractValueBehavior";
@@ -14,7 +14,6 @@ import Renderer from "component/Renderer";
 import argumentsBuilder from "function/argumentsBuilder";
 import Events from "const/EventsFields";
 import JSType from "const/JSType";
-import PropertyKeys from "const/PropertyKeys";
 import { Context, Stage } from "context/Context";
 import GlobalContextImpl from "context/GlobalContextImpl";
 import ForChannelContinuation from "continuation/ForChannelContinuation";
@@ -115,7 +114,6 @@ export {
 	FilterBuilder,
 	ForChannelContinuation,
 	Gettable,
-	Ids,
 	JSType,
 	LimitOffsetFilter,
 	Machine,
