--- packages/cydran/test/factory/Factories.spec.ts
@@ -1,12 +1,0 @@
-import { test, expect, jest } from "@jest/globals";
-import Factories from 'factory/Factories';
-import PropertiesImpl from "properties/PropertiesImpl";
-
-const WKPROPS: PropertiesImpl = new PropertiesImpl();
-
-
-test("importFactories", () => {
-  const wkSpy: Factories = jest.spyOn(Factories, 'importFactories');
-  Factories.importFactories(WKPROPS);
-  expect(wkSpy).toHaveBeenNthCalledWith(1, WKPROPS);
-});
