+++ packages/integration/test/testcase/digestion/no-components.spec.ts
@@ -1,11 +1,12 @@
-import { Ids, Stage, StageImpl } from "@cydran/cydran";
+import { create, Stage } from "@cydran/cydran";
+import { test, expect } from '@jest/globals';
 
-test.skip("Digestion - No behaviors", () => {
+test("Digestion - No behaviors", () => {
 	document.body.innerHTML = '<div></div>';
 
-	const stage: Stage = new StageImpl("body", {"cydran.logging.level": "WARN"});
+	const stage: Stage = create("body", {"cydran.logging.level": "WARN"});
 	stage.start();
 
 	expect(stage.isStarted()).toEqual(true);
-	expect(stage.getObject(Ids.STAGE)).not.toBeNull();
+	expect(stage.getContext().getObject("cydranStage")).not.toBeNull();
 });
